fix: move IPNet to ptypes, use it on network.Config
License: MIT
